+++ crates/ryo3-std/src/time/mod.rs
@@ -2,7 +2,7 @@ pub(crate) mod duration;
 pub(crate) mod functions;
 pub(crate) mod instant;
 pub use duration::PyDuration;
-pub use functions::{py_instant, sleep};
+pub use functions::{py_duration, py_instant, sleep};
 pub use instant::PyInstant;
 use pyo3::prelude::*;
 
@@ -10,6 +10,7 @@ pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<PyDuration>()?;
     m.add_class::<PyInstant>()?;
     m.add_function(wrap_pyfunction!(functions::py_instant, m)?)?;
+    m.add_function(wrap_pyfunction!(functions::py_duration, m)?)?;
     m.add_function(wrap_pyfunction!(functions::sleep, m)?)?;
     Ok(())
 }