+++ tests/fnv/test_fnv.py
@@ -10,10 +10,10 @@ from ._fnv_test_data import FNV_TEST_DATA
 
 
 def test_fnv_hasher_name() -> None:
-    assert ry.FnvHasher().__class__.__name__ == "FnvHasher"
+    assert ry.fnv1a().__class__.__name__ == "fnv1a"
     instance = ry.fnv1a(b"")
     assert instance.name == "fnv1a"
-    assert ry.FnvHasher.name == "fnv1a"
+    assert ry.fnv1a.name == "fnv1a"
     assert instance.name == "fnv1a"
 
 
@@ -56,31 +56,34 @@ def test_fnv_key_parse_err(bad_key: bytes | float | str) -> None:
 
 
 @pytest.mark.parametrize("data,expected", FNV_TEST_DATA)
-def test_fnv1a(data: bytes, expected: int) -> None:
-    fnvhash = ry.fnv1a(data)
-    int_digest = fnvhash.intdigest()
-    assert int_digest == expected
-    hex_str_expected = hex(expected)[2:]
+class TestFnv1a:
+    def test_fnv1a(self, data: bytes, expected: int) -> None:
+        fnvhash = ry.fnv1a(data)
+        int_digest = fnvhash.intdigest()
+        assert int_digest == expected
+        hex_str_expected = f"{expected:016x}"
+        hex_digest_str_og_hasher = fnvhash.hexdigest()
+        assert hex_digest_str_og_hasher == hex_str_expected
+        hex_digest_str = ry.fnv1a(data).hexdigest()
+        assert hex_digest_str == hex_str_expected
+        assert hex_digest_str == hex_digest_str.lower()
 
-    hex_digest_str_og_hasher = fnvhash.hexdigest()
-    assert hex_digest_str_og_hasher == hex_str_expected
-
-    hex_digest_str = ry.fnv1a(data).hexdigest()
-    assert hex_digest_str == hex_str_expected
-    assert hex_digest_str == hex_digest_str.lower()
+    def test_oneshot(self, data: bytes, expected: int) -> None:
+        int_digest = ry.fnv1a.oneshot(data)
+        assert int_digest == expected
 
 
 @pytest.mark.parametrize("data,expected", FNV_TEST_DATA)
 def test_fnv1a_hasher(data: bytes, expected: int) -> None:
-    thingy = ry.FnvHasher()
+    thingy = ry.fnv1a()
     thingy.update(data)
     assert thingy.intdigest() == expected
-    thingy_with_init = ry.FnvHasher(data)
+    thingy_with_init = ry.fnv1a(data)
     assert thingy_with_init.intdigest() == expected
 
 
 def test_copy_hasher() -> None:
-    thingy = ry.FnvHasher()
+    thingy = ry.fnv1a()
     thingy.update(b"abc")
     thingy_copy = thingy.copy()
     thingy_copy.update(b"def")